refactors rich data services and tests into separate package
