+++ hat/app/org/hatdex/hat/api/service/monitoring/EndpointSubscriberService.scala
@@ -25,7 +25,7 @@
 package org.hatdex.hat.api.service.monitoring
 
 import org.hatdex.hat.api.models._
-import org.hatdex.hat.api.service.JsonDataTransformer
+import org.hatdex.hat.api.service.richData.JsonDataTransformer
 import org.joda.time.DateTime
 import play.api.Logger
 import play.api.libs.json.Reads._
